fix: corrected points raised
